Move base intake url file to helpers
