+++ src/commands/sourcemaps/upload.ts
@@ -12,7 +12,7 @@ import {getMetricsLogger, MetricsLogger} from '../../helpers/metrics'
 import {upload, UploadStatus} from '../../helpers/upload'
 import {getRequestBuilder} from '../../helpers/utils'
 import {getRepositoryData, newSimpleGit, RepositoryData} from './common/git'
-import {getBaseSourcemapIntakeUrl} from './common/base-intake-url'
+import {getBaseSourcemapIntakeUrl} from '../../helpers/base-intake-url'
 import {Sourcemap} from './interfaces'
 import {
   renderCommandInfo,
